Problem: Flattening fails to include some deps
Flattening makes all your deps circular deps, which is supposed to
mean everything gets compiled in one big derivation. But it's not that
simple.

 - The reverseCircularDeps mechanism makes the assumption that the
   catalog resolver went through the circle and added all the deps as
   reverse circular deps to the top derivation.
 - The flatten mechanism makes the assumption that adding a dep to
   reverseCircularDeps is enough.

Solution: reverseCircularDeps: Don't add the src, add the srcs.

Each derivation calculates srcs from its own reverseCircularDeps, so
this makes srcs accumulate all the packages necessary. Unfortunately,
this means it adds redundant srcs, so we can't be as strict about
those any more. We'll silently skip unpacking a src with a name that
we already unpacked.

This is the quick and easy solution. A more correct solution would be
to resolve the catalog differently in flattening mode: Don't percolate
circular dependencies to the top and remove them as dependencies, just
break the circle at the bottom and be done.

Closes #156

+++ nix/racket2nix.rkt
@@ -97,16 +97,18 @@ mkRacketDerivation = suppliedAttrs: let racketDerivation = lib.makeOverridable (
     runHook preUnpack
     for unpackSrc in $srcs; do
       unpackName=$(stripSuffix $(stripHash $unpackSrc))
-      mkdir $unpackName
-      cd $unpackName
-      unpackFile $unpackSrc
-      cd -
-      unpackedFiles=( $unpackName/* )
-      if [ "''${unpackedFiles[*]}" = "$unpackName/$unpackName" ]; then
-        mv $unpackName _
-        chmod u+w _/$unpackName
-        mv _/$unpackName $unpackName
-        rmdir _
+      if ! [ -d $unpackName ]; then
+        mkdir $unpackName
+        cd $unpackName
+        unpackFile $unpackSrc
+        cd -
+        unpackedFiles=( $unpackName/* )
+        if [ "''${unpackedFiles[*]}" = "$unpackName/$unpackName" ]; then
+          mv $unpackName _
+          chmod u+w _/$unpackName
+          mv _/$unpackName $unpackName
+          rmdir _
+        fi
       fi
     done
     chmod u+w -R .
@@ -449,8 +451,8 @@ EOM
   (define srcs
     (cond
       [(pair? reverse-circular-build-inputs)
-       (define srcs-refs (string-join (map (lambda (s) (format "_~a.src" s)) reverse-circular-build-inputs)))
-       (format "~n  extraSrcs = [ ~a ];" srcs-refs)]
+       (define srcs-refs (string-join (map (lambda (s) (format "_~a.srcs" s)) reverse-circular-build-inputs) " ++ "))
+       (format "~n  extraSrcs = ~a;" srcs-refs)]
       [else ""]))
 
   (format derivation-template name (string-join (list src srcs) "")
